refactor(users): modify User ValueObject for enhanced domain logic.

+++ src/main/java/com/skillswap/platform/tutormatch/Users/Domain/Model/ValueObjects/Role.java
@@ -1,15 +1,15 @@
 package com.skillswap.platform.tutormatch.Users.Domain.Model.ValueObjects;
 
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
-public record Role(RoleType roleType, Integer tutorId) {
-    private static final AtomicInteger tutorIdCounter = new AtomicInteger(0);
+public record Role(RoleType roleType, Long tutorId) {
+    private static final AtomicLong tutorIdCounter = new AtomicLong(0);
 
     public Role(RoleType roleType) {
         this(roleType, roleType == RoleType.teacher ? generateTutorId() : null);
     }
 
-    private static int generateTutorId() {
+    private static Long generateTutorId() {
         return tutorIdCounter.incrementAndGet();
     }
 
@@ -17,7 +17,7 @@ public record Role(RoleType roleType, Integer tutorId) {
         return roleType;
     }
 
-    public Integer tutorId() {
+    public Long tutorId() {
         return tutorId;
     }
 }
